The key of a bucket can have different types
Like I said in the issue #19 the keys of buckets can be int and other types depending on the scheme of Elastic. I my case I did an aggregation to an id with the type integer. It is not possible to get the key of the bucket in the result without an PHP error. So I changed the type of the return value of the method "keys" (Bucket) from "string" to "mixed".